card read view split
